made xtp ready for c++17
