+++ CMakeLists.txt
@@ -26,7 +26,7 @@ endif()
 
 include(FeatureSummary)
 include(CheckCXXCompilerFlag)
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF) #...without compiler extensions like gnu++11
